requirements.txt
discord.py>=2.6.0
pytest>=8.2.0
python-dotenv>=1.0.0